+++ components/core/include/memfault/core/debug_log.h
@@ -6,13 +6,16 @@
 //! See License.txt for details
 //!
 //! @brief
+
 //! Log utility used within the memfault SDK. When enabled, logs will be emitted to help a user
-//! understand what is happening in the library. The Memfault SDK uses logs sparingly
-//! to better call out glaring configuration issues and errors.
+//! understand what is happening in the library.
+//!
+//! The Memfault SDK uses logs sparingly to better call out glaring configuration issues and
+//! runtime errors. It is recommended to enable all levels of Memfault logs.
 //!
 //! If your system does not have logging infrastructure in place, the subsystem can also be
-//! leveraged for logging within your platform. In that situation we suggest lightly wrapping
-//! the macros in your own file for future flexibility:
+//! leveraged for logging within your platform. In that situation, we suggest making your own log.h
+//! file for the platform and calling the Memfault macros from there:
 //!
 //! #include "memfault/core/debug_log.h"
 //! #define YOUR_PLATFORM_LOG_DEBUG(...) MEMAULT_LOG_DEBUG(__VA_ARGS__)
@@ -33,6 +36,21 @@ extern "C" {
   #define MEMFAULT_SDK_LOG_SAVE_DISABLE 0
 #endif
 
+// Allows for the MEMFAULT_LOG APIs to be re-mapped to another macro rather
+// than the  memfault_platform_log dependency function
+//
+// In this mode a user of the SDK will need to define the following macros:
+//   MEMFAULT_LOG_DEBUG(...)
+//   MEMFAULT_LOG_INFO(...)
+//   MEMFAULT_LOG_WARN(...)
+//   MEMFAULT_LOG_ERROR(...)
+//
+// And if the "demo" component is being used:
+//   MEMFAULT_LOG_RAW(...)
+#if !defined(MEMFAULT_PLATFORM_HAS_LOG_CONFIG)
+  #define MEMFAULT_PLATFORM_HAS_LOG_CONFIG 0
+#endif
+
 #if !MEMFAULT_SDK_LOG_SAVE_DISABLE
 // Note that this call will be a no-op if the system has not initialized the log module
 // by calling memfault_log_boot(). See ./log.h for more details.
@@ -41,13 +59,17 @@ extern "C" {
 #define MEMFAULT_SDK_LOG_SAVE(...)
 #endif
 
+#if MEMFAULT_PLATFORM_HAS_LOG_CONFIG
+  #include "memfault_platform_log_config.h"
+#else
+
 #define _MEMFAULT_LOG_IMPL(_level, ...)             \
   do {                                              \
     MEMFAULT_SDK_LOG_SAVE(_level, __VA_ARGS__);     \
     memfault_platform_log(_level, __VA_ARGS__);     \
   } while (0)
 
-#define MEMFAULT_LOG_DEBUG(...)                                             \
+#define MEMFAULT_LOG_DEBUG(...)                                         \
   _MEMFAULT_LOG_IMPL(kMemfaultPlatformLogLevel_Debug, __VA_ARGS__)
 
 #define MEMFAULT_LOG_INFO(...)                                              \
@@ -59,8 +81,11 @@ extern "C" {
 #define MEMFAULT_LOG_ERROR(...)                                             \
   _MEMFAULT_LOG_IMPL(kMemfaultPlatformLogLevel_Error, __VA_ARGS__)
 
-#define MEMFAULT_LOG_RAW(...)                                               \
+//! Only needs to be implemented when using demo component
+#define MEMFAULT_LOG_RAW(...)                   \
   memfault_platform_log_raw(__VA_ARGS__)
+
+#endif
 
 #ifdef __cplusplus
 }