+++ ports/esp_idf/memfault/common/memfault_platform_core.c
@@ -9,10 +9,12 @@
 #include "memfault/core/compiler.h"
 #include "memfault/core/debug_log.h"
 #include "memfault/core/event_storage.h"
+#include "memfault/core/data_packetizer.h"
 #include "memfault/core/platform/core.h"
 #include "memfault/core/reboot_tracking.h"
 #include "memfault/core/sdk_assert.h"
 #include "memfault/esp_port/cli.h"
+#include "memfault/metrics/metrics.h"
 
 #ifndef ESP_PLATFORM
 #  error "The port assumes the esp-idf is in use!"
@@ -24,12 +26,28 @@
 #include "freertos/semphr.h"
 #include "soc/cpu.h"
 #include "esp_system.h"
+#include "esp_timer.h"
 #include "esp_log.h"
 
 static uint8_t s_event_storage[CONFIG_MEMFAULT_EVENT_STORAGE_RAM_SIZE];
 static uint8_t s_log_buf_storage[CONFIG_MEMFAULT_LOG_STORAGE_RAM_SIZE];
 
 static __NOINIT_ATTR uint8_t s_reboot_tracking[MEMFAULT_REBOOT_TRACKING_REGION_SIZE];
+
+MEMFAULT_WEAK
+bool memfault_esp_port_data_available(void) {
+  return memfault_packetizer_data_available();
+}
+
+MEMFAULT_WEAK
+bool memfault_esp_port_get_chunk(void *buf, size_t *buf_len) {
+  return memfault_packetizer_get_chunk(buf, buf_len);
+}
+
+uint64_t memfault_platform_get_time_since_boot_ms(void) {
+  const int64_t time_since_boot_us = esp_timer_get_time();
+  return  (uint64_t) (time_since_boot_us / 1000) /* us per ms */;
+}
 
 bool memfault_arch_is_inside_isr(void) {
   return xPortInIsrContext();
@@ -99,17 +117,15 @@ void memfault_unlock(void) {
   xSemaphoreGiveRecursive(s_memfault_lock);
 }
 
-static int prv_memfault_log_wrapper(const char *fmt, va_list args) {
-  // flush to stdout
-  int rv = vprintf(fmt, args);
-  if (rv <= 0) {
-    return rv;
-  }
-
+// The esp32 uses vprintf() for dumping to console. The libc implementation used requires a lot of
+// stack space. We therefore prevent this function from being inlined so the log_buf allocation
+// does not wind up in that path.
+__attribute__((noinline))
+static int prv_copy_log_to_mflt_buffer(const char *fmt, va_list args) {
   // copy result into memfault log buffer collected as part of a coredump
   char log_buf[80];
   const size_t available_space = sizeof(log_buf);
-  rv = vsnprintf(log_buf, available_space, fmt, args);
+  const int rv = vsnprintf(log_buf, available_space, fmt, args);
 
   if (rv <= 0) {
     return -1;
@@ -121,6 +137,13 @@ static int prv_memfault_log_wrapper(const char *fmt, va_list args) {
   }
   memfault_log_save_preformatted(kMemfaultPlatformLogLevel_Info, log_buf, bytes_written);
   return rv;
+}
+
+static int prv_memfault_log_wrapper(const char *fmt, va_list args) {
+  prv_copy_log_to_mflt_buffer(fmt, args);
+
+  // flush to stdout
+  return vprintf(fmt, args);
 }
 
 // register an initialization routine that will be run from do_global_ctors()
@@ -136,6 +159,12 @@ static void __attribute__((constructor)) prv_memfault_boot(void) {
   const sMemfaultEventStorageImpl *evt_storage =
       memfault_events_storage_boot(s_event_storage, sizeof(s_event_storage));
   memfault_reboot_tracking_collect_reset_info(evt_storage);
+
+
+  sMemfaultMetricBootInfo boot_info = {
+    .unexpected_reboot_count = memfault_reboot_tracking_get_crash_count(),
+  };
+  memfault_metrics_boot(evt_storage, &boot_info);
 
   // register CLI for easily testing Memfault
   memfault_register_cli();